finalizado ejercicio 5, modificado ejercicio 4 para usar comparadores compatibles con jdk anteriores (no usando el metodo comparing(clase::getter) de la clase Comparator), definitivamente adoptado el cargador automatico para la fiaca de escribir 10 nombres por run

+++ Guia 10 - Colecciones/JavaColecciones/src/comparators/PeliculaDuracionComparator.java
@@ -7,7 +7,7 @@ public class PeliculaDuracionComparator implements PeliculaComparator{
 
     @Override
     public int compare(Pelicula pelicula1, Pelicula pelicula2) {
-        return Integer.compare((int)pelicula1.getDuracion(), (int)pelicula2.getDuracion());
+        return Double.compare(pelicula1.getDuracion(), pelicula2.getDuracion());
     }
     
 }
